Padding

+++ src/main.rs
@@ -24,7 +24,7 @@ pub fn prove_sha512(msg: &[u8]) -> Result<()> {
     type C = PoseidonGoldilocksConfig;
     type F = <C as GenericConfig<D>>::F;
     let mut builder = CircuitBuilder::<F, D>::new(CircuitConfig::wide_ecc_config());
-    let targets = make_circuits(&mut builder, len);
+    let targets = make_circuits(&mut builder, len as u128);
     let mut pw = PartialWitness::new();
 
     for i in 0..len {
